backend/requirements.txt
# CTD Stability Document Generator — Python dependencies

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy==2.0.35
asyncpg==0.30.0
alembic==1.13.0
psycopg2-binary==2.9.9

# File parsing
pdfplumber==0.11.0
python-docx==1.1.0
openpyxl==3.1.5
camelot-py[cv]==0.11.0
PyMuPDF==1.24.0
pytesseract==0.3.13

# Document generation
docxtpl==0.18.0
docx2pdf==0.1.8           # Windows only (requires MS Word)

# Task queue
celery==5.4.0
redis==5.1.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Storage
boto3==1.35.0             # For S3/MinIO
minio==7.2.0

# Utilities
httpx==0.27.0
Pillow==10.4.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
